Issue 42 data dump (#77)
* rule proposition

* fix - missing pattern

* comments

* design of rule in rql

* Rule has name in RQL

* code dumper

* Am I still normal ?

* rule formula evaluation

* cformat

* cformat

* limit ci

* work in progress

* dump manager proto

* Progress

* comments

* progress

* typo and bump circural buffer

* fixes

* work in progress

* fix on multiple set_capacity on circ buffer

* fix

* fix system action counting in compiler

* cleanup

* dumb assert

* working and testing

* testing update

* spdlog bump

* Massive rule mechanism update

* Presentation

* remove comments

* dot fix

+++ src/rdb/lib/facctxtsrc.cc
@@ -83,7 +83,7 @@ ssize_t textSourceAccessor::read(uint8_t *ptrData, const size_t position) {
             auto var = readFromFstream<unsigned>(myFile);
             payload->setItem(i + j, static_cast<uint8_t>(var));
           } else {
-            SPDLOG_ERROR("Unsupported type in text data source: {}", item.rtype);
+            SPDLOG_ERROR("Unsupported type in text data source: {}", static_cast<int>(item.rtype));
             assert(false && "read - Unsupported type in text data source");
             abort();
           }
